성민추가

+++ goldDog/src/main/java/com/goldDog/domain/HairstylistVO.java
@@ -9,14 +9,14 @@ public class HairstylistVO {
 	private int h_no;
 	private int m_no; 
 	private int a_no; 
-	private long h_price;
+	private int h_price;
 	private String h_self;
 	private String h_sel;
 	private String h_license;
 	private String h_businessNo;
 	private int h_best;
 	private int h_career;
-	private long h_salary;
+	private int h_salary;
 	private Timestamp h_date;
 	private double h_RAvg;
 	private int h_RTotal;
